pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
    "wheel",
]

[project]
name = "gdb_plus"
description = "Python library to automate gdb debugging"
version = "6.4.1"
authors = [{name = "Edoardo"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pwntools>=4.5.0",
    "capstone",
    "libdebug@git+https://github.com/Angelo942/libdebug.git@improve_detach",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: Unix",
]

[project.urls]
"Homepage" = "https://github.com/Angelo942/pydbg"
"Bug Tracker" = "https://github.com/Angelo942/pydbg/issues"